Move all UI components to web library plugin
