+++ packages/app/package.json
@@ -43,7 +43,6 @@
     "@bcgov/bc-sans": "^2.0.0",
     "@bcgov/design-tokens": "^3.0.0",
     "@internal/plugin-analytics-module-snowplow": "^0.1.0",
-    "@internal/plugin-policy-common": "^0.1.0",
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.3",
     "@material-ui/lab": "^4.0.0-alpha.61",
@@ -54,11 +53,7 @@
     "@snowplow/browser-tracker": "^3.15.0",
     "backstage-plugin-techdocs-addon-mermaid": "^0.13.0",
     "history": "^5.0.0",
-    "lodash-es": "^4.17.21",
-    "luxon": "^3.5.0",
-    "parse-github-url": "1.0.3",
     "react": "^18.0.2",
-    "victory": "^37.3.6",
     "react-dom": "^18.0.2",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
