+++ packages/app/src/components/catalog/EntityPage.tsx
@@ -56,18 +56,16 @@ import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
 import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
 import { Mermaid } from 'backstage-plugin-techdocs-addon-mermaid';
 
-import { EntityPoliciesCard } from './EntityPoliciesCard';
 import {
+  EntityPoliciesCard,
   ComponentSecurityAlertsContent,
   SystemSecurityAlertsContent,
-} from './SecurityAlerts';
-import { PolicyComplianceCard } from './PolicyComplianceCard';
-import {
+  PolicyComplianceCard,
   PolicyToolsCard,
   PolicyMostCompliantCard,
   PolicyLeastCompliantCard,
-} from './PolicyOverviewCards';
-import { PolicyScopeCard } from './PolicyScopeCard';
+  PolicyScopeCard,
+} from 'backstage-plugin-policy-react';
 
 const techdocsContent = (
   <EntityTechdocsContent>
