+++ plugins/policy-react/package.json
@@ -25,8 +25,17 @@
     "postpack": "backstage-cli package postpack"
   },
   "dependencies": {
+    "@backstage/core-components": "^0.15.1",
     "@backstage/core-plugin-api": "^1.9.3",
-    "@material-ui/core": "^4.9.13"
+    "@backstage/plugin-catalog": "^1.24.0",
+    "@backstage/plugin-catalog-react": "^1.14.0",
+    "@internal/plugin-policy-common": "^0.1.0",
+    "@material-ui/core": "^4.9.13",
+    "@material-ui/icons": "^4.11.3",
+    "luxon": "^3.5.0",
+    "lodash-es": "^4.17.21",
+    "parse-github-url": "1.0.3",
+    "victory": "^37.3.6"
   },
   "peerDependencies": {
     "react": "^16.13.1 || ^17.0.0 || ^18.0.0"
