+++ plugins/policy-react/src/components/EntityPoliciesCard/EntityPoliciesCard.tsx
