--- plugins/policy-react/src/components/ExampleComponent/index.ts
@@ -1,2 +1,0 @@
-export { ExampleComponent } from './ExampleComponent';
-export type { ExampleComponentProps } from './ExampleComponent';
