+++ plugins/policy-react/src/components/PolicyScopeCard/index.tsx
