+++ plugins/policy-react/src/components/index.ts
@@ -2,4 +2,8 @@
 // The index file in ./components/ is typically responsible for selecting
 // which components are public API and should be exported from the package.
 
-export * from './ExampleComponent';
+export * from './EntityPoliciesCard';
+export * from './PolicyComplianceCard';
+export * from './PolicyOverviewCards';
+export * from './PolicyScopeCard';
+export * from './SecurityAlerts';
