--- plugins/policy-react/src/hooks/index.ts
@@ -1,5 +1,0 @@
-/***/
-// The index file in ./hooks/ is typically responsible for selecting
-// which hooks are public API and should be exported from the package.
-
-export * from './useExample';
