--- plugins/policy-react/src/hooks/useExample/index.ts
@@ -1,1 +1,0 @@
-export { useExample } from './useExample';
